az204/02_security/key-vault/sdk/requirements.txt
python-dotenv
# provides AAD token auth support
azure-identity
azure-keyvault-secrets
azure-keyvault-keys
azure-keyvault-certificates
